@@ -2,12 +2,10 @@ require 'test_helper'
 require 'database_cleaner'
 
 class SchedulableTest < ActiveSupport::TestCase
-  puts "HELLO"
-  
   DatabaseCleaner.clean_with(:truncation)
   DatabaseCleaner.start
-  
-  event = FactoryGirl.create(:event)
+
+  event = FactoryBot.create(:event)
   puts event.name
 
   test "truth" do
