@@ -19,11 +19,10 @@ module Dummy
     # The default locale is :en and all translations from config/locales/*.rb,yml are auto loaded.
     # config.i18n.load_path += Dir[Rails.root.join('my', 'locales', '*.{rb,yml}').to_s]
     config.i18n.default_locale = :de
-    
+
     config.assets.paths << Rails.root.join('vendor', 'assets', 'components')
-    
+
     config.action_view.embed_authenticity_token_in_remote_forms = true
-    
   end
 end
 
